requirements.txt
langchain-core
langchain-community
langchain-openai
python-dotenv
openai
requests
pyyaml
pygithub
ipython
pytest
